feat:新增取消、刪除、拒絕、接受邀約的websocket通知

+++ src/controllers/invitationController.ts
@@ -2,7 +2,7 @@ import { type NextFunction, type Request, type Response } from "express"
 import { type LoginResData } from "@/types/login"
 import { Invitation } from "@/models/invitation"
 import { BeInvitation } from "@/models/beInvitation"
-import { Profile } from "@/models/profile"
+import { Profile, type IPersonalInfo } from "@/models/profile"
 import { Collection } from "@/models/collection"
 import appErrorHandler from "@/utils/appErrorHandler"
 import appSuccessHandler from "@/utils/appSuccessHandler"
@@ -34,6 +34,10 @@ interface ParsedInvitation {
 // import { eraseProperty } from "@/utils/responseDataHandler"
 const postInvitation = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   const { userId } = req.user as LoginResData
+  if (!userId) {
+    appErrorHandler(400, "缺少使用者Id", next)
+    return
+  }
   const { invitedUserId, message } = req.body
   if (!invitedUserId) {
     appErrorHandler(400, "缺少邀請使用者Id", next)
@@ -74,7 +78,7 @@ const postInvitation = async (req: Request, res: Response, next: NextFunction): 
       return
     }
   }
-  const invitation = await Invitation.create({ userId, invitedUserId, message })
+  const [invitation, profileWithUser] = await Promise.all([Invitation.create({ userId, invitedUserId, message }), Profile.findOne({ userId }).select("nickNameDetails")])
   const isNotificationCreated = await createNotification(userId, invitedUserId, message, 1)
   if (!invitation || !isNotificationCreated) {
     appErrorHandler(400, "邀請失敗", next)
@@ -86,7 +90,10 @@ const postInvitation = async (req: Request, res: Response, next: NextFunction): 
     if (!beInvitation) {
       appErrorHandler(400, "邀請失敗", next)
     } else {
-      sendNotification({ title: message.title, content: message.content })
+      const { nickNameDetails } = profileWithUser as IPersonalInfo
+      const { userId } = req.user as LoginResData
+      sendNotification({ title: message.title, content: message.content, nickNameDetails, userId })
+      console.log({ title: message.title, content: message.content, nickNameDetails, userId })
       appSuccessHandler(201, "邀請成功", invitation, res)
     }
   }
@@ -152,22 +159,29 @@ const getInvitationById = async (req: Request, res: Response, next: NextFunction
 const cancelInvitation = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   const { id } = req.params
   const { userId } = req.user as LoginResData
-  const invitation = await Invitation.findByIdAndUpdate(id, { status: "cancel" }, { new: true })
-
+  const [invitation, profileWithUser] = await Promise.all([Invitation.findByIdAndUpdate(id, { status: "cancel" }, { new: true }), Profile.findOne({ userId }).select("nickNameDetails")])
   const beInvitation = await BeInvitation.findOneAndUpdate({ userId, invitationId: id }, { status: "cancel" }, { new: true })
   if (!invitation || !beInvitation) {
     appErrorHandler(404, "No invitation found", next)
   } else {
+    const { nickNameDetails } = profileWithUser as IPersonalInfo
+    const { userId } = req.user as LoginResData
+    sendNotification({ title: "取消邀約", content: `${nickNameDetails.nickName}已取消邀約`, nickNameDetails, userId })
     appSuccessHandler(200, "取消邀請成功", invitation, res)
   }
 }
 const deleteInvitation = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   const { id } = req.params
-  const invitation = await Invitation.findByIdAndDelete(id)
+  const { userId } = req.user as LoginResData
+  // const invitation = await Invitation.findByIdAndDelete(id)
+  const [invitation, profileWithUser] = await Promise.all([Invitation.findByIdAndDelete(id), Profile.findOne({ userId }).select("nickNameDetails")])
   const beInvitation = await BeInvitation.findOneAndUpdate({ invitationId: id }, { status: "cancel" }, { new: true })
   if (!invitation || !beInvitation) {
     appErrorHandler(404, "No invitation found", next)
   } else {
+    const { nickNameDetails } = profileWithUser as IPersonalInfo
+    const { userId } = req.user as LoginResData
+    sendNotification({ title: "取消邀約", content: `${nickNameDetails.nickName}已取消邀約`, nickNameDetails, userId })
     appSuccessHandler(200, "刪除成功", invitation, res)
   }
 }
